reposync: Prepared for dnf plugin.
