+++ backend/satellite_tools/repo_plugins/yum_zypper_src.py
@@ -49,6 +49,7 @@ from salt.utils.versions import LooseVersion
 from uyuni.common import checksum, fileutils
 from spacewalk.common import rhnLog
 from spacewalk.satellite_tools.repo_plugins import ContentPackage, CACHE_DIR
+from spacewalk.satellite_tools.repo_plugins.yum_src_common import RepoMDError, UpdateNotice, UpdateNoticeException
 from spacewalk.satellite_tools.download import get_proxies
 from spacewalk.common.rhnConfig import CFG, initCFG
 from spacewalk.common.suseLib import get_proxy
@@ -72,6 +73,8 @@ REPOSYNC_EXTRA_HTTP_HEADERS_CONF = '/etc/rhn/spacewalk-repo-sync/extra_headers.c
 
 RPM_PUBKEY_VERSION_RELEASE_RE = re.compile(r'^gpg-pubkey-([0-9a-fA-F]+)-([0-9a-fA-F]+)')
 
+APACHE_USER = 'wwwrun'
+APACHE_GROUP = 'www'
 
 class ZyppoSync:
     """
@@ -157,9 +160,9 @@ class ZypperRepo:
            self.urls[0] += '/'
        # Make sure root paths are created
        if not os.path.isdir(self.root):
-           fileutils.makedirs(self.root, user='wwwrun', group='www')
+           fileutils.makedirs(self.root, user=APACHE_USER, group=APACHE_GROUP)
        if not os.path.isdir(self.pkgdir):
-           fileutils.makedirs(self.pkgdir, user='wwwrun', group='www')
+           fileutils.makedirs(self.pkgdir, user=APACHE_USER, group=APACHE_GROUP)
        self.is_configured = False
        self.includepkgs = []
        self.exclude = []
@@ -204,186 +207,9 @@ class RawSolvablePackage:
         return epoch, version, release
 
 
-class RepoMDError(Exception):
-    """ An exception thrown when not RepoMD is found. """
-    pass
-
-
 class SolvFileNotFound(Exception):
     """ An exception thrown when not Solv file is found. """
     pass
-
-
-class UpdateNoticeException(Exception):
-    """ An exception thrown for bad UpdateNotice data. """
-    pass
-
-
-class UpdateNotice(object):
-    """
-    Simplified UpdateNotice class implementation
-    https://github.com/rpm-software-management/yum/blob/master/yum/update_md.py
-
-    A single update notice (for instance, a security fix).
-    """
-    def __init__(self, elem=None, repoid=None, vlogger=None):
-        self._md = {
-            'from'             : '',
-            'type'             : '',
-            'title'            : '',
-            'release'          : '',
-            'status'           : '',
-            'version'          : '',
-            'pushcount'        : '',
-            'update_id'        : '',
-            'issued'           : '',
-            'updated'          : '',
-            'description'      : '',
-            'rights'           : '',
-            'severity'         : '',
-            'summary'          : '',
-            'solution'         : '',
-            'references'       : [],
-            'pkglist'          : [],
-            'reboot_suggested' : False,
-            'restart_suggested' : False
-        }
-
-        if elem is not None:
-            self._parse(elem)
-
-    def __getitem__(self, item):
-        """ Allows scriptable metadata access (ie: un['update_id']). """
-        if type(item) is int:
-            return sorted(self._md)[item]
-        ret = self._md.get(item)
-        if ret == '':
-            ret = None
-        return ret
-
-    def __setitem__(self, item, val):
-        self._md[item] = val
-
-    def _parse(self, elem):
-        """
-        Parse an update element::
-            <!ELEMENT update (id, synopsis?, issued, updated,
-                              references, description, rights?,
-                              severity?, summary?, solution?, pkglist)>
-                <!ATTLIST update type (errata|security) "errata">
-                <!ATTLIST update status (final|testing) "final">
-                <!ATTLIST update version CDATA #REQUIRED>
-                <!ATTLIST update from CDATA #REQUIRED>
-        """
-        if elem.tag == 'update':
-            for attrib in ('from', 'type', 'status', 'version'):
-                self._md[attrib] = elem.attrib.get(attrib)
-            for child in elem:
-                if child.tag == 'id':
-                    if not child.text:
-                        raise UpdateNoticeException("No id element found")
-                    self._md['update_id'] = child.text
-                elif child.tag == 'pushcount':
-                    self._md['pushcount'] = child.text
-                elif child.tag == 'issued':
-                    self._md['issued'] = child.attrib.get('date')
-                elif child.tag == 'updated':
-                    self._md['updated'] = child.attrib.get('date')
-                elif child.tag == 'references':
-                    self._parse_references(child)
-                elif child.tag == 'description':
-                    self._md['description'] = child.text
-                elif child.tag == 'rights':
-                    self._md['rights'] = child.text
-                elif child.tag == 'severity':
-                    self._md[child.tag] = child.text
-                elif child.tag == 'summary':
-                    self._md['summary'] = child.text
-                elif child.tag == 'solution':
-                    self._md['solution'] = child.text
-                elif child.tag == 'pkglist':
-                    self._parse_pkglist(child)
-                elif child.tag == 'title':
-                    self._md['title'] = child.text
-                elif child.tag == 'release':
-                    self._md['release'] = child.text
-        else:
-            raise UpdateNoticeException('No update element found')
-
-    def _parse_references(self, elem):
-        """
-        Parse the update references::
-            <!ELEMENT references (reference*)>
-            <!ELEMENT reference>
-                <!ATTLIST reference href CDATA #REQUIRED>
-                <!ATTLIST reference type (self|other|cve|bugzilla) "self">
-                <!ATTLIST reference id CDATA #IMPLIED>
-                <!ATTLIST reference title CDATA #IMPLIED>
-        """
-        for reference in elem:
-            if reference.tag == 'reference':
-                data = {}
-                for refattrib in ('id', 'href', 'type', 'title'):
-                    data[refattrib] = reference.attrib.get(refattrib)
-                self._md['references'].append(data)
-            else:
-                raise UpdateNoticeException('No reference element found')
-
-    def _parse_pkglist(self, elem):
-        """
-        Parse the package list::
-            <!ELEMENT pkglist (collection+)>
-            <!ELEMENT collection (name?, package+)>
-                <!ATTLIST collection short CDATA #IMPLIED>
-                <!ATTLIST collection name CDATA #IMPLIED>
-            <!ELEMENT name (#PCDATA)>
-        """
-        for collection in elem:
-            data = { 'packages' : [] }
-            if 'short' in collection.attrib:
-                data['short'] = collection.attrib.get('short')
-            for item in collection:
-                if item.tag == 'name':
-                    data['name'] = item.text
-                elif item.tag == 'package':
-                    data['packages'].append(self._parse_package(item))
-            self._md['pkglist'].append(data)
-
-    def _parse_package(self, elem):
-        """
-        Parse an individual package::
-            <!ELEMENT package (filename, sum, reboot_suggested, restart_suggested)>
-                <!ATTLIST package name CDATA #REQUIRED>
-                <!ATTLIST package version CDATA #REQUIRED>
-                <!ATTLIST package release CDATA #REQUIRED>
-                <!ATTLIST package arch CDATA #REQUIRED>
-                <!ATTLIST package epoch CDATA #REQUIRED>
-                <!ATTLIST package src CDATA #REQUIRED>
-            <!ELEMENT reboot_suggested (#PCDATA)>
-            <!ELEMENT restart_suggested (#PCDATA)>
-            <!ELEMENT filename (#PCDATA)>
-            <!ELEMENT sum (#PCDATA)>
-                <!ATTLIST sum type (md5|sha1) "sha1">
-        """
-        package = {}
-        for pkgfield in ('arch', 'epoch', 'name', 'version', 'release', 'src'):
-            package[pkgfield] = elem.attrib.get(pkgfield)
-
-        #  Bad epoch and arch data is the most common (missed) screwups.
-        # Deal with bad epoch data.
-        if not package['epoch'] or package['epoch'][0] not in '0123456789':
-            package['epoch'] = None
-
-        for child in elem:
-            if child.tag == 'filename':
-                package['filename'] = child.text
-            elif child.tag == 'sum':
-                package['sum'] = (child.attrib.get('type'), child.text)
-            elif child.tag == 'reboot_suggested':
-                self._md['reboot_suggested'] = True
-            elif child.tag == 'restart_suggested':
-                self._md['restart_suggested'] = True
-        return package
 
 
 class ContentSource:
@@ -1142,7 +968,6 @@ type=rpm-md
         if not os.path.exists(target_dir):
             os.makedirs(target_dir, int('0755', 8))
 
-        params['authtoken'] = self.authtoken
         params['urls'] = self.repo.urls
         params['relative_path'] = relative_path
         params['authtoken'] = self.authtoken
